Address review feedback: credential guards, shared helper, exit codes
P1: Add requireProvidersOrFail() guard to cross-provider tests so they
fail fast with a clear message in CI instead of timing out when
MINIMAX_API_KEY or GLM_API_KEY are absent (per CLAUDE.md fail-not-skip rule).

P2: Add process.exit(1) to investigation script when Bug 2 is confirmed
(or inconclusive), making it CI-integratable as a regression test.

P3: Extract waitForSystemInit() to a shared helper
(packages/daemon/tests/helpers/sdk-message-helpers.ts) to eliminate
duplication between rpc-model-switching.test.ts and
bug2-model-switch-investigation.ts.
